# src/frontend/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CMS",
        "description": "CMS endpoints",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "CMS - Admin API server"
        }
    ],
    "paths": {
...
        "/tenants": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "summary": "Get paginated list of tenants",
                "operationId": "getTenants",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of records to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Number of records to skip",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Tenant"
                                            }
                                        },
                                        "paging": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "limit": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "offset": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tenant"
                ],
                "summary": "Create a new tenant",
                "operationId": "createTenant",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TenantInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tenant created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tenant"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "properties": {
                                                "slug": {
                                                    "type": "string",
                                                    "example": "The slug field is required."
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "The name field is required."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tenants/all": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "summary": "Get all tenants",
                "operationId": "getAllTenants",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tenant"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/tenants/domain/{domain}": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "summary": "Get tenant by domain",
                "operationId": "getTenantByDomain",
                "parameters": [
                    {
                        "name": "domain",
                        "in": "path",
                        "description": "Tenant domain",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Tenant"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tenant not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/tenants/org/{orgId}": {
            "get": {
                "tags": [
                    "Tenant"
                ],
                "summary": "Get tenants by Organization ID",
                "operationId": "getTenantsByOrgId",
                "parameters": [
                    {
                        "name": "orgId",
                        "in": "path",
                        "description": "Organization ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Tenant"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Internal Server Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Tenant": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Tenant Name"
                    },
                    "domain": {
                        "type": "string",
                        "example": "https://example-tenant.com"
                    },
                    "config": {
                        "type": "object",
                        "example": {
                            "theme": "dark",
                            "locale": "en"
                        }
                    },
                    "status": {
                        "type": "string",
                        "example": "draft"
                    },
                    "org_id": {
                        "type": "integer",
                        "example": 42
                    },
                    "created_at": {
                        "properties": {
                            "date": {
                                "type": "string",
                                "example": "2025-06-08 02:13:02.000000"
                            },
                            "timezone_type": {
                                "type": "integer",
                                "example": 3
                            },
                            "timezone": {
                                "type": "string",
                                "example": "UTC"
                            }
                        },
                        "type": "object"
                    },
                    "updated_at": {
                        "properties": {
                            "date": {
                                "type": "string",
                                "example": "2025-06-08 02:13:02.000000"
                            },
                            "timezone_type": {
                                "type": "integer",
                                "example": 3
                            },
                            "timezone": {
                                "type": "string",
                                "example": "UTC"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "TenantInput": {
                "required": [
                    "domain",
                    "name",
                    "status",
                    "config",
                    "org_id"
                ],
                "properties": {
                    "domain": {
                        "description": "Tenant domain (valid URL). Must be at least 3 characters and unique.",
                        "type": "string",
                        "example": "https://example-tenant.com"
                    },
                    "name": {
                        "description": "Display name for the tenant. Must be at least 3 characters.",
                        "type": "string",
                        "example": "New Tenant"
                    },
                    "status": {
                        "description": "Status of the tenant.",
                        "type": "string",
                        "enum": [
                            "draft",
                            "published",
                            "archived"
                        ],
                        "example": "draft"
                    },
                    "config": {
                        "description": "Configuration in JSON format.",
                        "type": "object",
                        "example": {
                            "theme": "dark",
                            "locale": "en"
                        }
                    },
                    "org_id": {
                        "description": "ID of the associated organization.",
                        "type": "integer",
                        "example": 42
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}